Add files via upload
Adding this logo size to use it as a static link for our internal and community Jitsi. If the server is updated it will stick, as before it would need to be set after each update. We didn't have this logo in a small enough size, but now we do :)